rename folder entitiesData
