+++ src/main/kotlin/data/entitiesData/LogData.kt
@@ -1,4 +1,4 @@
-package data.dbEntities
+package data.entitiesData
 
 import java.time.LocalDateTime
 import java.util.UUID
