+++ src/main/kotlin/data/entitiesData/TaskData.kt
@@ -1,10 +1,12 @@
-package data.dbEntities
+package data.entitiesData
 
 import java.util.UUID
 
-data class ProjectData(
+data class TaskData(
     val id: UUID,
     val title: String,
     val description: String,
+    val stateId: UUID,
+    val projectId: UUID,
     val isDeleted: Boolean
 )
