+++ src/main/kotlin/data/entitiesData/TaskStateData.kt
@@ -1,12 +1,11 @@
-package data.dbEntities
+package data.entitiesData
 
 import java.util.UUID
 
-data class TaskData(
+data class TaskStateData(
     val id: UUID,
     val title: String,
     val description: String,
-    val stateId: UUID,
     val projectId: UUID,
     val isDeleted: Boolean
 )
